+++ mpv.net/WPF/AboutWindow.xaml.cs
@@ -12,8 +12,9 @@ namespace mpvnet
         public AboutWindow()
         {
             InitializeComponent();
-            Version.Text = $"mpv.net Version {System.Windows.Forms.Application.ProductVersion} ({File.GetLastWriteTime(System.Windows.Forms.Application.ExecutablePath).ToShortDateString()})";
-            mpvVersion.Text = $"{core.get_property_string("mpv-version")} ({File.GetLastWriteTime(Folder.Startup + "mpv-1.dll").ToShortDateString()})";
+            ContentBlock.Text= "Copyright (C) 2017-2020 Frank Skare (stax76)\n" +
+                $"mpv.net {System.Windows.Forms.Application.ProductVersion} ({File.GetLastWriteTime(System.Windows.Forms.Application.ExecutablePath).ToShortDateString()})\n" +
+                $"{core.get_property_string("mpv-version")} ({File.GetLastWriteTime(Folder.Startup + "mpv-1.dll").ToShortDateString()})\nffmpeg {core.get_property_string("ffmpeg-version")}\nMIT License";
         }
 
         protected override void OnPreviewKeyDown(KeyEventArgs e) => Close();